refactor: isolate standalone key setup and service composition
